+++ nuztf/ampel/urls.py
@@ -1,0 +1,5 @@
+API_BASEURL = "https://ampel.zeuthen.desy.de"
+API_ZTF_ARCHIVE_URL = API_BASEURL + "/api/ztf/archive/v3"
+API_CATALOGMATCH_URL = API_BASEURL + "/api/catalogmatch"
+
+MAX_N_PIX = 1000
