+++ nuztf/ampel/utils.py
@@ -1,0 +1,43 @@
+"""
+Utility functions for Ampel API
+"""
+from nuztf.credentials import load_credentials
+
+def merge_alerts(alert_list: list) -> list:
+    """ """
+    merged_list = []
+    keys = list(set([x["objectId"] for x in alert_list]))
+
+    for objectid in keys:
+        alerts = [x for x in alert_list if x["objectId"] == objectid]
+        if len(alerts) == 1:
+            merged_list.append(alerts[0])
+        else:
+            jds = [x["candidate"]["jd"] for x in alerts]
+            order = [jds.index(x) for x in sorted(jds)[::-1]]
+            latest = alerts[jds.index(max(jds))]
+            latest["candidate"]["jdstarthist"] = min(
+                [x["candidate"]["jdstarthist"] for x in alerts]
+            )
+
+            for index in order[1:]:
+                x = alerts[index]
+
+                # Merge previous detections
+
+                for prv in x["prv_candidates"] + [x["candidate"]]:
+                    if prv not in latest["prv_candidates"]:
+                        latest["prv_candidates"] = [prv] + latest["prv_candidates"]
+
+            merged_list.append(latest)
+
+    return merged_list
+
+def get_ampel_token() -> str:
+    """
+    Function to get the ampel token from the environment
+    """
+
+    # Load credentials from environment
+    _, ampel_api_archive_token = load_credentials("ampel_api_archive_token")
+    return ampel_api_archive_token
