+++ nuztf/credentials.py
@@ -11,7 +11,9 @@ from ztfquery import io
 
 
 def load_credentials(name: str, token_based: bool = False):
-    """ZTFquery wrapper for loading credentials."""
+    """
+    ZTFquery wrapper for loading credentials.
+    """
     return io._load_id_(name, token_based=token_based)
 
 
@@ -95,3 +97,15 @@ try:
 
 except KeyError:
     logging.info("No Token for Fritz API found in environment" "Assume it is set.")
+
+try:
+    with warnings.catch_warnings():
+        warnings.filterwarnings("ignore", category=UserWarning)
+        io.set_account(
+            which="kowalski", token=os.environ["KOWALSKI_API_TOKEN"], token_based=True,
+            force=True
+        )
+        logging.info('Set up "kowalski" credentials')
+
+except KeyError:
+    logging.info("No Token for Kowalski API found in environment" "Assume it is set.")
