+++ pyproject.toml
@@ -45,6 +45,9 @@ slackclient = {version = "^2.9.4", optional = true}
 gunicorn = {version = ">=20.1,<24.0", optional = true}
 slackeventsapi = {version = "^3.0.1", optional = true}
 typer = {version = "^0.15.1", optional = true}
+penquins = "^2.4.2"
+dotenv = "^0.9.9"
+pydantic = "^2.10.6"
 
 
 [tool.poetry.scripts]
